phylorug: Visualize Clade Recovery and Support Across Phylogenetic Trees

Generates rug plot visualizations directly on tree nodes to illustrate clade recovery and support from multiple phylogenetic and phylogenomic analyses. See vignette("phylorug") for a detailed guide.

Core functions

read_trees

Read tree files from disk

translate_tips

Rename specimen codes to species names

check_taxa

Diagnose taxon consistency across trees

prune_to_shared

Prune trees to shared taxa

node_presence_matrix

Make presence/support matrix

plot_phylorug

Draw the rug plot on a reference tree

See also

Author

Maintainer: Md Rifath Ahamed rifath.ahamed@helsinki.fi (ORCID)

Authors:

Md Rifath Ahamed rifath.ahamed@helsinki.fi (ORCID)

Sergei Tarasov sergei.tarasov@helsinki.fi (ORCID)

Other contributors:

J. Salvador Arias salvador.ariasbecerra@helsinki.fi (ORCID) [contributor]